Rename competable_app_mocks.dart
